basic seed packing

+++ kardashev-seed/src/config.rs
@@ -2,5 +2,5 @@ use serde::Deserialize;
 
 #[derive(Debug, Deserialize)]
 pub struct Config {
-    pub local_bubble: Option<f64>,
+    pub local_bubble: Option<f32>,
 }
